mcp/rag-mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-mcp-server"
version = "0.1.0"
description = "RAG MCP Server for Claude Code — Semantic search over PDFs, code, and documents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Rubrum95", email = "jtvr90@gmail.com"},
]
dependencies = [
    "mcp>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "PyMuPDF>=1.23.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]

[project.scripts]
rag-mcp-server = "rag_mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_mcp_server"]
